Projeto pronto

+++ AspNetMvcRoles/Controllers/RegrasController.cs
@@ -1,22 +1,25 @@
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using AspNetMvcRoles.Data;
 using AspNetMvcRoles.Models;
+using Microsoft.AspNetCore.Identity;
+using System;
+using Microsoft.AspNetCore.Authorization;
 
 namespace AspNetMvcRoles.Controllers
 {
+    [Authorize]
     public class RegrasController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly RoleManager<IdentityRole> _roleManager;
 
-        public RegrasController(ApplicationDbContext context)
+        public RegrasController(ApplicationDbContext context, RoleManager<IdentityRole> roleManager)
         {
             _context = context;
+            _roleManager = roleManager;
         }
 
         // GET: Regras
@@ -50,16 +53,32 @@ namespace AspNetMvcRoles.Controllers
         }
 
         // POST: Regras/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("IdRegra,Name")] Regras regras)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(regras);
-                await _context.SaveChangesAsync();
+                IdentityRole role = new IdentityRole();
+                role.Name = regras.Name;
+                role.NormalizedName = regras.Name;
+                role.ConcurrencyStamp = DateTime.Now.ToString("yyyyMMddHHmmssffff");
+                var roleResult = await _roleManager.CreateAsync(role);
+
+                if (!roleResult.Succeeded)
+                {
+                    foreach (var erro in roleResult.Errors)
+                    {
+                        ModelState.AddModelError("Name", erro.ToString());
+
+                        return CreatedAtAction("GetRegras", new { id = regras.IdRegra }, regras);
+                    }
+                }
+                else
+                {
+                    _context.Add(regras);
+                    await _context.SaveChangesAsync();
+                }
                 return RedirectToAction(nameof(Index));
             }
             return View(regras);
@@ -82,8 +101,6 @@ namespace AspNetMvcRoles.Controllers
         }
 
         // POST: Regras/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("IdRegra,Name")] Regras regras)
@@ -99,10 +116,11 @@ namespace AspNetMvcRoles.Controllers
                 {
                     _context.Update(regras);
                     await _context.SaveChangesAsync();
+                    UpdateRegra(id, regras.Name);
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!RegrasExists(regras.IdRegra))
+                    if (!RegrasExists(id))
                     {
                         return NotFound();
                     }
@@ -111,9 +129,31 @@ namespace AspNetMvcRoles.Controllers
                         throw;
                     }
                 }
+
                 return RedirectToAction(nameof(Index));
             }
             return View(regras);
+        }
+
+        public async void UpdateRegra(int id, string nome)
+        {
+            
+                var consult_regras = await _context.Regras.FindAsync(id);
+                String nome_regra = nome;
+                var role = await _roleManager.FindByNameAsync(consult_regras.Name);
+                role.Name = nome_regra;
+                role.NormalizedName = nome_regra;
+                role.ConcurrencyStamp = DateTime.Now.ToString("yyyyMMddHHmmssffff");
+                var roleResult = await _roleManager.UpdateAsync(role);
+
+                if (!roleResult.Succeeded)
+                {
+                    foreach (var erro in roleResult.Errors)
+                    {
+                        ModelState.AddModelError("Name", erro.ToString());
+                    }
+                }
+
         }
 
         // GET: Regras/Delete/5
@@ -140,8 +180,27 @@ namespace AspNetMvcRoles.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var regras = await _context.Regras.FindAsync(id);
-            _context.Regras.Remove(regras);
-            await _context.SaveChangesAsync();
+
+            var role = await _roleManager.FindByNameAsync(regras.Name);
+            var roleResult = await _roleManager.DeleteAsync(
+                    await _roleManager.FindByIdAsync(role.Id)
+                );
+
+            if (!roleResult.Succeeded)
+            {
+                foreach (var erro in roleResult.Errors)
+                {
+                    ModelState.AddModelError("Name", erro.ToString());
+
+                    return CreatedAtAction("GetRegras", new { id = regras.IdRegra }, regras);
+                }
+            }
+            else
+            {
+                _context.Regras.Remove(regras);
+                await _context.SaveChangesAsync();
+            }
+
             return RedirectToAction(nameof(Index));
         }
